Api endpoint: companies done. Auth endpoints done. API configurations done.

+++ backend/src/companies/companies.service.ts
@@ -5,6 +5,7 @@ import { GetCompaniesFilterDto } from './dto/get-companies-filter-dto';
 import { CompanyRepository } from './company.repository';
 import { CreateCompanyDto } from './dto/create-company-dto';
 import { CompanyStatus } from './company-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class CompaniesService {
@@ -13,52 +14,47 @@ export class CompaniesService {
     private readonly companyRepository: CompanyRepository,
   ) {}
 
-  async createCompany(createCompanyDto: CreateCompanyDto): Promise<Company> {
+  async createCompany(
+    createCompanyDto: CreateCompanyDto, 
+    loggedUser: User
+    ): Promise<Company> {
     // Using repository
-    return this.companyRepository.createCompany(createCompanyDto);
+    return this.companyRepository.createCompany(createCompanyDto, loggedUser);
   }
 
-  async getCompanies(filterDto: GetCompaniesFilterDto): Promise<Company[]> {
-    return await this.companyRepository.getCompanies(filterDto);
+  async getCompanies(
+    filterDto: GetCompaniesFilterDto,
+    loggedUser: User
+    ): Promise<Company[]> {
+    return await this.companyRepository.getCompanies(filterDto, loggedUser);
   }
 
-  async getCompanyById(id: number): Promise<Company> {
-    const r = await this.companyRepository.findOneBy({ id });
+  async getCompanyById(
+    id: number,
+    loggedUser: User
+    ): Promise<Company> {
+    const r = await this.companyRepository.findOne({ where: { id, userId: loggedUser.id } });
     // console.log(r);
-    if (!r) throw new NotFoundException(`Company with ID ${id} not found.`);
+    if (!r) throw new NotFoundException();
+    // if (!r) throw new NotFoundException(`Company with ID ${id} not found.`);
     return r;
   }
 
-  async deleteCompany(id: number): Promise<void> {
-    const r = await this.companyRepository.delete(id);
+  async deleteCompany(
+    id: number,
+    loggedUser: User
+    ): Promise<void> {
+    const r = await this.companyRepository.delete({ id, userId: loggedUser.id });
     // console.log(r);
     if ( r.affected === 0 ) throw new NotFoundException(`Company with ID ${id} not found.`);
   }
 
-  // async filter(filter: GetCompaniesFilterDto): Promise<Company[]> {
-    
-  //   const { status, search } = filter;
-
-  //   // console.log(filter);
-
-  //   let companies = await this.getCompanies();
-
-  //   if (status)
-  //     companies = companies.filter((company) => company.active === status);
-
-  //   if (search) {
-  //     companies = companies.filter(
-  //       (company) =>
-  //         company.name.includes(search) ||
-  //         company.registryNumber.includes(search),
-  //     );
-  //   }
-
-  //   return companies;
-  // }
-
-  async updateCompanyStatus(id: number, status: CompanyStatus): Promise<Company> {
-    const r = await this.getCompanyById(id);
+  async updateCompanyStatus(
+    id: number, 
+    status: CompanyStatus,
+    loggedUser: User
+    ): Promise<Company> {
+    const r = await this.getCompanyById(id, loggedUser);
     r.active = status;
     await r.save();
     return r;
